Added Numberpad-Support
Added Numberpad-Support for level 2-6.
Changed shortcut for toggling bypass-mode from shift + pause to shift + scroll-lock (because keycode 69 was the keycode for pause and numlock)

Known Bugs:
Strangely enough the middle row (456) don't work.
The € for shift + 5 only works when numberLock is activated.
The operator characters print the same boring characters in every level. e.g.  Mod4 + '-' ≠ ∓ (it should though)
There is a bidirectional relationship between the numberkeys and the non-printing control keys. So the arrow key and so on produce numbers. Why??

Unknown Bugs:
Probably some more